refactor: move WeaponProfileFields beside the other shared form fields
components/admin holds self-contained modals and nothing else; shared form
components live outside it. AvailabilityPicker is the direct precedent -- a
Necromunda-specific field used by both equipment modals and two customise
components, sitting in components/ui alongside the generic primitives. So ui
is not reserved for shadcn wrappers, and this belongs there rather than
introducing the first cross-import inside components/admin.

It also puts the component where components/customise/custom-weapon-profiles
can reach it. That file is a third copy of the same form, out of scope here but
the obvious next consumer.

File move and two import lines; no code change.
