+++ components/admin/admin-create-equipment.tsx
@@ -10,7 +10,7 @@ import { toast } from 'sonner';
 import { WeaponProfileInput, EquipmentAvailability, GangAdjustedCost, emptyWeaponProfile } from "@/types/equipment";
 import { EditionSelect, useEditions, editionSlugOf } from '@/components/edition-select';
 import { hasLethalityStatline } from '@/types/edition';
-import { WeaponProfileFields } from '@/components/admin/weapon-profile-fields';
+import { WeaponProfileFields } from '@/components/ui/weapon-profile-fields';
 import { HiX } from "react-icons/hi";
 import { LuTrash2 } from 'react-icons/lu'
