+++ components/ui/weapon-profile-fields.tsx
